requirements.txt
# KISWARM v1.1 — Pinned Production Dependencies
# Install: pip install -r requirements.txt

# Core AI / Memory
ollama>=0.1.7
mem0>=0.1.0
qdrant-client>=1.7.0

# Vector / ML utilities
chromadb>=0.4.0
tiktoken>=0.5.0
numpy>=1.24.0

# Web Framework (Tool Proxy)
flask>=3.0.0
flask-cors>=4.0.0

# HTTP
requests>=2.31.0

# Monitoring / UI
rich>=13.0.0
psutil>=5.9.0

# File watching
watchdog>=3.0.0

# AI providers (optional)
openai>=1.0.0

# Type support
typing-extensions>=4.8.0

# ── Dev / Test dependencies ────────────────────────────────────────────────
# Install with: pip install -r requirements-dev.txt

# ── v2.1 Sentinel Bridge ───────────────────────────────────────────────────
aiohttp>=3.9.0
sentence-transformers>=2.7.0
pytest-asyncio>=0.21.0
